# api-swagger.yaml
openapi: 3.0.1
info:
  title: Superheroes
  description: Superheroes API
  version: 0.0.1
servers:
  - url: http://localhost:8081/api/v1
    description: Local server
tags:
  - name: Superheroes
paths:
  /superheroes:
    get:
      tags:
        - Superheroes
      summary: Get all superheroes
      operationId: getAllSuperheroes
      parameters:
        - name: name.contains
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Superheroes'
  /superheroes/{id}:
    get:
      tags:
        - Superheroes
      summary: Get one superhero
      operationId: getOneSuperhero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Superheroes'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SUPERHERO_NOT_FOUND-1000-404'
    put:
      tags:
        - Superheroes
      summary: Modify superhero
      operationId: modifySuperhero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Superheroes'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SUPERHERO_NOT_FOUND-1000-404'
    delete:
      tags:
        - Superheroes
      summary: Delete superhero
      operationId: deleteSuperhero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        204:
          description: No content
          content: {}
components:
  schemas:
    Superheroes:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Batman

    SUPERHERO_NOT_FOUND-1000-404:
      type: object
      properties:
        errorCode:
          type: string
          default: 1000
        message:
          type: string
          default: Superhero not found